src/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2
idna==3.4
mailchimp-marketing==3.0.80
Pillow==9.5.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
django-recaptcha==3.0.0
plotly==5.15.0
pandas==2.0.2
reportlab==4.0.4